settings page done with delete and and update profile
